Image queue procedure library
image_queue.hdpl is a procedure library that creates a FIFO queue for
monochrome images with a specified image and stack size.  This can be
used for applying an operator to a defined image stack, such as a simple
average, a running average or a running sum.
image_queue_to_mean_image.hdev demonstrates using the queue procedures
to create a running average.